Manager: Retry snap installs
The way this worked previously was that it would exit out early if it
wasn't a DNS exit code (100). So for snap installs, this would mean a
retry wouldn't happen as that was a 1.

The code is changed so that we have a set of high-order functions that
allow you to express what we want to achieve from calling the packaging
manager install.

Unfortunately we had to make some breaking changes and bump to v2.