Manager: Re-work the runner
The running of a retry command now attempts to split the work into 2
places. The running of the command and if the error is a fatal error.
With this in place we don't need to check for retrying locally and we
can ask each individual package manager if they can retry.

The runner previously had way to much knowledge and pushed apt, yum,
snap and zypper to all behave the same. This isn't the case and not how
each package manager works.

This simple change, isn't so simple because the tests patch the hell out
of things and calls commands locally on each system, which is confusing.

The changes here try and straighten it all out, but could do with
re-write of the tests as they're observed testing, rather than expected
conditional testing.