Remove unused functions + drop series before focal
We wish to remove the concept of a series from the juju extended
codebase. Series is something that juju/packaging relies on quite
heavily

We can change this by removing a number of unused functions/methods
which take in series as parameters.

Also, since we no longer support series before focal, we can take out a
lot of the special cases and make the PackingingConfigurers series
agnostic